Add an SSH domain for remote development at work
I can see how it can replace tmux. So far, a workflow I settled on is to
create a new workspace, attach to the remote domain, detach when
finished. When reattached, the windows are placed in a workspace with
the same name as when the domain was detached. Overall very similar to
the tmux workflow, with a benefit of considering whether I actually need
to leave a session hanging when done.

Having to have a compatible WezTerm executable on the remote host is not
as bad, as it comes as an AppImage and can be put wherever and pointed
to by a parameter of the domain.