Add a combo to switch between layer and option
Now pressing ,. at the same time toggles a layer in which the tap dance
for the colored key is switched from enter/layer_1/option to
enter/option/layer_1, and the same for the second colored key.

The default is great for typing. The added variant is great for
meta-tasks like switching and organizing windows.